requirements.txt
python-telegram-bot==13.15
gspread==6.1.2
google-auth==2.35.0
